Revert "ajustes configurações"
This reverts commit 60d92bbf27141ca0d50b6f9d8607a2c50c939095.

+++ src/tema/services/tema.service.ts
@@ -1,7 +1,7 @@
 import { HttpException, HttpStatus, Injectable } from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
 import { DeleteResult, ILike, Repository } from "typeorm";
-import { Tema } from "../entities/tema.entity";
+import { Tema } from "./../entities/tema.entity";
 
 @Injectable()
 export class TemaService {
@@ -19,6 +19,7 @@ export class TemaService {
     }
 
     async findById(id: number): Promise<Tema> {
+
         let tema = await this.temaRepository.findOne({
             where: {
                 id
@@ -34,7 +35,7 @@ export class TemaService {
         return tema;
     }
 
-    async findAllByDescricao(descricao: string): Promise<Tema[]> {
+    async findByDescricao(descricao: string): Promise<Tema[]> {
         return await this.temaRepository.find({
             where: {
                 descricao: ILike(`%${descricao}%`)
@@ -51,14 +52,20 @@ export class TemaService {
 
     async update(tema: Tema): Promise<Tema> {
 
-        await this.findById(tema.id);
+        let buscaTema = await this.findById(tema.id);
+
+        if (!buscaTema || !tema.id)
+            throw new HttpException('Tema não encontrado!', HttpStatus.NOT_FOUND);
 
         return await this.temaRepository.save(tema);
     }
 
     async delete(id: number): Promise<DeleteResult> {
 
-        await this.findById(id);
+        let buscaTema = await this.findById(id);
+
+        if (!buscaTema)
+            throw new HttpException('Tema não encontrado!', HttpStatus.NOT_FOUND);
 
         return await this.temaRepository.delete(id);
